add NO_ARCH_NATIVE environment variable
 NO_ARCH_NATIVE environment variable control -march=native optimization.
User must export NO_ARCH_NATIVE=1 to disable optimization.
With this option enable (by default), falcON compiled on a recent architecture could crash at runtime on an older machine if using the same binary. It can happen on cluster and also on docker which are supposed to run on multiple machines.